moved step under compute
